refactor(serdes): move parts into their own files
