+++ packages/qwik/src/core/client/dom-container.ts
@@ -7,13 +7,8 @@ import { QError, qError } from '../shared/error/error';
 import { ERROR_CONTEXT, isRecoverable } from '../shared/error/error-handling';
 import type { QRL } from '../shared/qrl/qrl.public';
 import { _SharedContainer } from '../shared/shared-container';
-import {
-  getObjectById,
-  inflateQRL,
-  parseQRL,
-  preprocessState,
-  wrapDeserializerProxy,
-} from '../shared/shared-serialization';
+import { getObjectById, inflateQRL, parseQRL, preprocessState } from '../shared/serdes/index';
+import { wrapDeserializerProxy } from '../shared/serdes/deser-proxy';
 import { QContainerValue, type HostElement, type ObjToProxyMap } from '../shared/types';
 import { EMPTY_ARRAY } from '../shared/utils/flyweight';
 import {
